Add more xml docs.

+++ tests/NetworkPrimitives.Switching.Tests/Common/EndianSwapTests.cs
@@ -10,7 +10,7 @@ namespace NetworkPrimitives.Tests.Common
     {
         [Test]
         [TestCase((uint)0x00112233, (uint)0x33221100)]
-        [TestCase((uint)0xFF116600, (uint)0x006611FF)]
+        [TestCase(0xFF116600, (uint)0x006611FF)]
         public void TestEndianSwap(uint input, uint expected)
         {
             Assert.AreEqual(expected, input.SwapEndian());
